Explicit conversion to avoid compiler warning
Small change just to avoid compiler warning

+++ Utilities/StrFmt.h
@@ -85,7 +85,7 @@ namespace fmt{
 		{
 			std::vector<char> buffptr(length);
 			int printlen = snprintf(buffptr.data(), length, fmt.c_str(), std::forward<Args>(parameters)...);
-			if (printlen >= 0 && printlen < length)
+			if (printlen >= 0 && (size_t)printlen < length)
 			{
 				str = string(buffptr.data(), printlen);
 				break;
